chore(eslint-plugin-cdk): tests for rule no-qualified-construct (#12809)
Add missing tests

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
