+++ tools/eslint-plugin-cdk/test/rules/fixtures.test.ts
@@ -1,0 +1,63 @@
+import { ESLint } from 'eslint';
+import * as fs from 'fs-extra';
+import * as path from 'path';
+
+const rulesDirPlugin = require('eslint-plugin-rulesdir');
+rulesDirPlugin.RULES_DIR = path.join(__dirname, '../../lib/rules');
+
+const linter = new ESLint({
+  baseConfig: {
+    parser: '@typescript-eslint/parser',
+    plugins: ['rulesdir'],
+    rules: {
+      quotes: [ 'error', 'single', { avoidEscape: true }],
+      'rulesdir/no-core-construct': [ 'error' ],
+      'rulesdir/no-qualified-construct': [ 'error' ],
+    }
+  },
+  rulePaths: [
+    path.join(__dirname, '../../lib/rules'),
+  ],
+  fix: true,
+});
+
+const outputRoot = path.join(process.cwd(), '.test-output');
+fs.mkdirpSync(outputRoot);
+
+const fixturesRoot = path.join(__dirname, 'fixtures');
+
+fs.readdirSync(fixturesRoot).filter(f => fs.lstatSync(path.join(fixturesRoot, f)).isDirectory()).forEach(d => {
+  describe(d, () => {
+    const outputDir = path.join(outputRoot, d);
+    fs.mkdirpSync(outputDir);
+
+    const fixturesDir = path.join(fixturesRoot, d);
+    const fixtureFiles = fs.readdirSync(fixturesDir).filter(f => f.endsWith('.ts') && !f.endsWith('.expected.ts'));
+
+    fixtureFiles.forEach(f => {
+      test(f, async (done) => {
+        const actualFile = await lintAndFix(path.join(fixturesDir, f), outputDir);
+        const expectedFile = path.join(fixturesDir, `${path.basename(f, '.ts')}.expected.ts`);
+        if (!fs.existsSync(expectedFile)) {
+          done.fail(`Expected file not found. Generated output at ${actualFile}`);
+        }
+        const actual = await fs.readFile(actualFile, { encoding: 'utf8' });
+        const expected = await fs.readFile(expectedFile, { encoding: 'utf8' });
+        if (actual !== expected) {
+          done.fail(`Linted file did not match expectations. Expected: ${expectedFile}. Actual: ${actualFile}`);
+        }
+        done();
+      });
+    });
+  });
+});
+
+async function lintAndFix(file: string, outputDir: string) {
+  const newPath = path.join(outputDir, path.basename(file))
+  let result = await linter.lintFiles(file);
+  await ESLint.outputFixes(result.map(r => {
+    r.filePath = newPath;
+    return r;
+  }));
+  return newPath;
+}
