+++ tools/eslint-plugin-cdk/test/rules/fixtures/no-qualified-construct/qualified-heritage.ts
@@ -1,0 +1,4 @@
+import * as cdk from '@aws-cdk/core';
+
+class MyConstruct extends cdk.Construct {
+}
