+++ tools/eslint-plugin-cdk/test/rules/fixtures/no-qualified-construct/qualified-usage.expected.ts
@@ -1,0 +1,4 @@
+import * as cdk from '@aws-cdk/core';
+import * as constructs from 'constructs';
+
+let x: constructs.Construct;
