+++ tools/eslint-plugin-cdk/test/rules/fixtures/no-qualified-construct/qualified-usage.ts
@@ -1,0 +1,3 @@
+import * as cdk from '@aws-cdk/core';
+
+let x: cdk.Construct;
